requirements.txt
attrs==23.2.0
Babel==2.10.3
blinker==1.7.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
colorama==0.4.6
configobj==5.0.8
cryptography==41.0.7
defer==1.0.4
httplib2==0.20.4
idna==3.6
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
pexpect==4.9.0
pillow==10.2.0
ptyprocess==0.7.0
pycairo==1.25.1
Pygments==2.17.2
PyJWT==2.7.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-dateutil==2.8.2
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
setuptools==68.1.2
six==1.16.0
typing_extensions==4.10.0
urllib3==2.0.7
wadllib==1.3.6
wheel==0.42.0
xdg==5
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
redis
django-redis
channels>=4.1
channels_redis>=4.1
graphene-django>=3.2
